+++ pkg/attributesdecorator/decorator.go
@@ -39,3 +39,11 @@ func DecorateResourceAttributes[T Resource](collection ResourceCollection[T], at
 		}
 	}
 }
+
+func DecorateResourceAttributesByFunction[T Resource](collection ResourceCollection[T], fn func(resourceAttributes pcommon.Map)) {
+	for i := 0; i < collection.Len(); i++ {
+		resource := collection.At(i).Resource()
+		resourceAttributes := resource.Attributes()
+		fn(resourceAttributes)
+	}
+}